requirement.txt
langchain
python-dotenv
langchain_community
langchain_text_splitters
chromadb
langchain_chroma
langchain_core
langchain_huggingface
langchain_groq
langgraph
langchain
sentence-transformers
transformers
pypdf
python-docx
fastapi
unicorn
python-multipart